cleanup log route
