+++ app/views/admin/neurons/logs/_changes.html.slim
