+++ app/views/admin/neurons/logs/show.html.slim
